Make ProductOrVariation logic reusable in future
