+++ Yosemite/Yosemite/PointOfSale/Items/PointOfSaleBarcodeScanService.swift
@@ -1,7 +1,6 @@
 import Foundation
 import protocol Networking.ProductsRemoteProtocol
 import class Networking.ProductsRemote
-import class WooFoundation.CurrencyFormatter
 import class WooFoundation.CurrencySettings
 import class Networking.AlamofireNetwork
 
@@ -19,14 +18,15 @@ public enum PointOfSaleBarcodeScanError: Error, Equatable {
 public final class PointOfSaleBarcodeScanService: PointOfSaleBarcodeScanServiceProtocol {
     private let productsRemote: ProductsRemoteProtocol
     private let siteID: Int64
-    private let itemMapper: PointOfSaleItemMapper
+    private let itemResolver: POSProductOrVariationResolver
 
     init (siteID: Int64,
           productsRemote: ProductsRemoteProtocol,
           currencySettings: CurrencySettings) {
         self.siteID = siteID
         self.productsRemote = productsRemote
-        self.itemMapper = PointOfSaleItemMapper(currencySettings: currencySettings)
+        self.itemResolver = POSProductOrVariationResolver(productsRemote: productsRemote,
+                                                          currencySettings: currencySettings)
     }
 
     public convenience init(siteID: Int64,
@@ -45,66 +45,9 @@ public final class PointOfSaleBarcodeScanService: PointOfSaleBarcodeScanServiceP
         do {
             let productOrVariation = try await productsRemote.fetchPOSProductByGlobalUniqueIdentifier(for: siteID,
                                                                                                       globalUniqueID: barcode)
-            return try await itemForScannedBarcodeProduct(productOrVariation)
+            return try await itemResolver.itemForProductOrVariation(productOrVariation)
         } catch {
             throw PointOfSaleBarcodeScanError.unknown
         }
-    }
-
-    private func itemForScannedBarcodeProduct(_ productOrVariation: POSProduct) async throws -> POSItem {
-        let items: [POSItem]
-        switch productOrVariation.productType {
-        case .simple:
-            let product = productOrVariation
-            items = itemMapper.mapProductsToPOSItems(products: [product])
-        case .custom("variation"):
-            let variation = try productOrVariation.toProductVariation()
-            let variableProduct = try await parentProductForVariation(variation)
-            items = itemMapper.mapVariationsToPOSItems(variations: [variation], parentProduct: variableProduct)
-        default:
-            throw PointOfSaleBarcodeScanError.unknown
-        }
-
-        guard let item = items.first else {
-            throw PointOfSaleBarcodeScanError.unknown
-        }
-        return item
-    }
-
-    private func parentProductForVariation(_ variation: POSProductVariation) async throws -> POSVariableParentProduct {
-        guard variation.productID != 0 else {
-            throw PointOfSaleBarcodeScanError.noParentProductForVariation
-        }
-
-        let parentProduct = try await productsRemote.fetchPOSProduct(for: siteID, productID: variation.productID)
-        let mappedProducts = itemMapper.mapProductsToPOSItems(products: [parentProduct])
-
-        guard let mappedProduct = mappedProducts.first,
-              case .variableParentProduct(let variableProduct) = mappedProduct else {
-            throw PointOfSaleBarcodeScanError.variationCouldNotBeConverted
-        }
-        return variableProduct
-    }
-}
-
-
-private extension POSProduct {
-    func toProductVariation() throws -> POSProductVariation {
-        POSProductVariation(
-            siteID: siteID,
-            productID: parentID,
-            productVariationID: productID,
-            attributes: try attributes.compactMap { try $0.toProductVariationAttribute() },
-            image: images.first,
-            sku: sku,
-            globalUniqueID: globalUniqueID,
-            price: price,
-            regularPrice: regularPrice,
-            salePrice: salePrice,
-            onSale: onSale,
-            downloadable: downloadable,
-            manageStock: manageStock,
-            stockQuantity: stockQuantity,
-            stockStatusKey: stockStatusKey)
     }
 }
